correct map center radius for firefox + new osm
The way that the center radius was being calculated for both chrome + ff
used object properties that either were there before and were removed
with new open street maps or were possibly how google calculated it.
Now, the shape of the object is detected and the appropriate calculation
for detecting the center of the map search is used.